fix(schemas): allow fractional HF trendingScore
HuggingFace's trending-models API now returns trendingScore as a
float (e.g. 183.5); HFModelRow declared it as strict int, so pydantic
rejected every row and the hf_trending source went degraded.

+++ src/aipulse/schemas.py
@@ -42,7 +42,7 @@ class HFModelRow(BaseModel):
     downloads: int = 0
     downloadsAllTime: int | None = None
     likes: int = 0
-    trendingScore: int = 0
+    trendingScore: float = 0
     pipeline_tag: str | None = None
     library_name: str | None = None
